build: split file for build target information
